requirements.txt
asgiref==3.8.1
Django
psycopg2-binary==2.9.10
sqlparse==0.5.3
djangorestframework-simplejwt
djangorestframework
drf-yasg